ai_models/requirements.txt
# Core AI & Vision
ultralytics       # YOLOv8/v9/v10 model support
opencv-python     # Video and camera frame processing
torch             # PyTorch deep learning framework
torchvision       # Image transforms and pretrained model utilities

# Tracking & Counting
supervision       # ByteTrack multi-object tracking and line zone counting

# Data & Math
numpy             # Numerical array operations
scipy             # .mat file parsing (ShanghaiTech ground truth)
Pillow            # Image loading and preprocessing

# Optional — Logging & Visualization
matplotlib        # Density map and metric visualization (debugging)
pandas            # Structured result logging (optional)
